+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To Be Released
 
+* feat(events) add database backup events
+
 ## 8.7.0
 
 * feat(applications): create a new method to set the project